style(cfd-1d): Inline format-args + winnow `map_or` -> `is_some_and` (ratchet: 26 uninlined_format_args + 20 auto-fixable) (#312)
First CFDRS-CFD1D-LINT-001 ratchet decrement. `cargo clippy --fix`
applied to cfd-1d --all-targets (lint floor: touched code meets the
raised floor, brownfield debt uses the ratchet baseline per
engineering_gates lint-floor policy).

Baseline:
- Pre-decrement cfd-1d pedantic warnings: 54 (auto-fixer reached 26
  `uninlined_format_args` + 20 other auto-fixable lints --
  `unnecessary_map_or` (`map_or(false, ...)` -> `is_some_and(...)`),
  `useless_conversion` (`iter().copied().collect()` -> `to_vec()`),
  minor `.into_iter()`/`.into()` cleanups).
- Post-decrement: 8 warnings remain, all manual-only categories:
  3 `result_large_err` (architectural error-type redesign needed),
  1 `very_complex_type` (type-factor requirement),
  1 `empty_line_after_doc_comments`, 3 doc-test wrap warnings.
- Net delta: -46 warnings (-85%), +42 / -93 lines across 12 files.

Sites remediated (file:line -- category):
- crates/cfd-1d/src/domain/network/builder/venturi_coefficients.rs:
  multiple test asserts -- `uninlined_format_args`
- crates/cfd-1d/src/physics/cell_separation/cascade_junction/mod.rs:
  test asserts -- `uninlined_format_args`
- crates/cfd-1d/src/physics/cell_separation/fahraeus_lindqvist.rs:
  12 test asserts -- `uninlined_format_args`
- crates/cfd-1d/src/physics/cell_separation/plasma_skimming.rs:
  5 test asserts -- `uninlined_format_args`
- crates/cfd-1d/src/physics/cell_separation/rouleaux_aggregation.rs:
  4 test asserts -- `uninlined_format_args`
- crates/cfd-1d/src/physics/droplet_regime.rs: 1 assert --
  `uninlined_format_args`
- crates/cfd-1d/src/physics/resistance/models/entrance.rs:
  2 asserts -- `uninlined_format_args`
- crates/cfd-1d/src/physics/vascular/murrays_law/mod.rs:
  2 asserts -- `uninlined_format_args`
- crates/cfd-1d/tests/adversarial_tests.rs:
  pressures Vec construction -- `useless_conversion`->`to_vec()`
- crates/cfd-1d/tests/blueprint_solve_trace.rs:
  6 helper sites -- `unnecessary_map_or`->`is_some_and(...)`
  AND `uninlined_format_args`
- crates/cfd-1d/tests/resistance_model_validation.rs:
  1 doc -- `uninlined_format_args`

All fixes are mechanical clippy auto-suggestions -- no semantic change
to assertion logic, control flow, or test oracles. Verified:
cargo nextest run -p cfd-1d: 728/728 PASS, 3 skipped, 0 timeouts,
0 regressions (5.234s suite).

Residual 8-warning baseline (next ratchet decrement candidates) parked
as peer-architectural decisions (error-type redesign, type-factor,
doc-comment line cleanup); out-of-scope for this style-pass.

Refs: backlog.md#CFDRS-CFD1D-LINT-001

+++ crates/cfd-1d/src/domain/network/builder/venturi_coefficients.rs
@@ -137,11 +137,11 @@ mod tests {
         )?;
         let total_length = metadata.throat_length_m + conv_len + diff_len;
 
-        let inlet_d_t = f64::from(inlet_d);
-        let throat_d_t = f64::from(throat_d);
-        let outlet_d_t = f64::from(outlet_d);
-        let throat_len_t = f64::from(metadata.throat_length_m);
-        let total_len_t = f64::from(total_length);
+        let inlet_d_t = inlet_d;
+        let throat_d_t = throat_d;
+        let outlet_d_t = outlet_d;
+        let throat_len_t = metadata.throat_length_m;
+        let total_len_t = total_length;
 
         let manual =
             VenturiModel::new(inlet_d_t, throat_d_t, outlet_d_t, throat_len_t, total_len_t)
